Extract window width logic into a custom hook
